Attempt AdaBelief and madgrad optimizers (requires a separate install)
Add eps as an argument for the adabelief optimizer.  Could use it elsewhere as well
